+++ src/MetaheuristicsPlatform/Algorithms/BigBangBigCrunch/BigBangBigCrunchReferences.cs
@@ -1,0 +1,14 @@
+using MetaheuristicsPlatform.Classification;
+
+namespace MetaheuristicsPlatform.Algorithms.BigBangBigCrunch;
+
+public static class BigBangBigCrunchReferences
+{
+    public static ScientificReference ErolEksin2006 { get; } =
+        new(
+            "Erol & Eksin",
+            2006,
+            "A new optimization method: Big Bang-Big Crunch",
+            "Advances in Engineering Software 37(2), 106-111",
+            "10.1016/j.advengsoft.2005.04.005");
+}